+++ pkg/objectpacker/objectpacker_test.go
